Move api to pkg
To group Go sources into top level dir
