+++ pkg/operator/controllers/utils.go
@@ -5,14 +5,13 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/ccremer/clustercode/pkg/api/v1alpha1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/utils/pointer"
-
-	"github.com/ccremer/clustercode/api/v1alpha1"
 )
 
 type (
